Dzielenie

+++ src/lab6/P2.java
@@ -1,15 +1,9 @@
 package lab6;
 
-/*
- * 
- *  Funkcja Dodawania
- *   
- * */
-
 public class P2 {
 
-	static float dodawanie(float a, float b) {
-		float wynik = a + b;
+	static int dodawanie(int a, int b) {
+		int wynik = a + b;
 		return wynik;
 	}
